+++ assets/src/scripts/components/dailyValueHydrograph/index.js
@@ -8,7 +8,9 @@ import {Actions} from '../../store';
 import {drawErrorAlert, drawInfoAlert} from '../../d3-rendering/alerts';
 import {drawLoadingIndicator} from '../../d3-rendering/loading-indicator';
 
+
 import {drawTimeSeriesGraph} from './time-series-graph';
+import {drawTooltipCursorSlider} from './tooltip';
 
 const TEMP_TIME_SERIES_ID = '36307c899ac14d2eac6956b1bf5ceb69';
 
@@ -50,6 +52,6 @@ export const attachToNode = function (store,
         .call(link(store, function(container, showElem) {
             container.attr('hidden', showElem ? null : true);
         }, hasCurrentObservationsTimeSeries))
-        .call(drawTimeSeriesGraph, store);
-
+        .call(drawTimeSeriesGraph, store)
+        .call(drawTooltipCursorSlider, store);
 };